Configuration package was removed; created options class for user service; custom exceptions were deleted as unnecessary; other minor changes and renames
