+++ Identity/Identity.Application/DataObjects/UserServiceOptions.cs
@@ -1,0 +1,10 @@
+namespace Identity.Application.DataObjects;
+
+public class UserServiceOptions
+{
+    public string Pepper { get; set; }
+
+    public string Key { get; set; }
+    
+    public int Expires { get; set; }
+}
