--- Identity/Identity.Application/Exceptions/PermissionDeniedException.cs
@@ -1,6 +1,0 @@
-namespace Identity.Application.Exceptions;
-
-public class PermissionDeniedException : Exception
-{
-    public PermissionDeniedException(string message = "") : base(message) {}
-}
